Khoa

+++ portal_job_FN/portal_job_FN/Repositories/EFPostJobRepository.cs
@@ -17,18 +17,37 @@ namespace portal_job_FN.Repositories
             var applicationDbContext = await _context.post_Jobs
                  .Include(b => b.job_Location)
                 .Include(b => b.major)
+                .Include(b => b.applyJobs)
                 .Include(b => b.experience)
                 .Include(b => b.applicationUser)
                 .ToListAsync();
 
             return await _context.post_Jobs.ToListAsync();
         }
+
+        public async Task<IEnumerable<PostJob>> GetAllByCompanyIdAsync(string id)
+        {
+            //bao gồm danh mục, nếu không có sẽ ko ra danh mục
+            var applicationDbContext = await _context.post_Jobs
+                 .Include(b => b.job_Location)
+                .Include(b => b.major)
+                .Include(b => b.applyJobs)
+                .Include(b => b.experience)
+                .Include(b => b.applicationUser)
+                .Where(b => b.applicationUser.Id == id)
+                .ToListAsync();
+
+            return applicationDbContext;
+        }
+
+
         public async Task<PostJob> GetByIdAsync(int id)
         {
             var applicationDbContext = await _context.post_Jobs
                 .Include(b => b.job_Location)
                 .Include(b => b.major)
                 .Include(b => b.experience)
+                .Include(b => b.applyJobs)
                 .Include(b => b.applicationUser)
                 .ToListAsync();
             return await _context.post_Jobs.FindAsync(id);
@@ -46,11 +65,14 @@ namespace portal_job_FN.Repositories
             await _context.SaveChangesAsync();
         }
 
-        public async Task? DeleteAsync(int id)
+        public async Task DeleteAsync(int id)
         {
             var post_Job = await _context.post_Jobs.FindAsync(id);
-            _context.post_Jobs.Remove(post_Job);
-            await _context.SaveChangesAsync();
+            if (post_Job != null)
+            {
+                _context.post_Jobs.Remove(post_Job);
+                await _context.SaveChangesAsync();
+            }
         }
